+++ imx-core-sdk-kotlin-jvm/src/main/kotlin/com/immutable/sdk/model/SignableTokenData.kt
@@ -1,7 +1,11 @@
 package com.immutable.sdk.model
 
+import com.immutable.sdk.api.model.SignableToken
 import com.squareup.moshi.Json
 
+/**
+ * An object representing the `data` field of [SignableToken].
+ */
 data class SignableTokenData(
     @Json(name = "token_address")
     val tokenAddress: String? = null,